server/requirements.txt
flask-socketio
itsdangerous
